requirements.txt
fastapi[all]
motor
pymongo
python-decouple